requirements.txt
absl-py==0.15.0
acm-sdk-python==0.4.11
alembic==1.4.1
aniso8601==9.0.1
ann==0.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopep8==1.5.6
azure-core==1.10.0
azure-storage-blob==12.7.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==4.0.0
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
bson==0.5.10
cached-property==1.5.2
cachetools==4.1.1
certifi==2021.5.30
cffi==1.14.6
cftime==1.3.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
config==0.5.0.post0
configparser==5.0.2
confuse==1.4.0
contextlib2==0.6.0.post1
contextvars==2.4
convertdate==2.3.2
coverage==5.5
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.24
cytoolz==0.11.0
dask==2021.3.0
databricks-cli==0.14.1
dataclasses==0.8
DateTime==4.3
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.2
distributed==2021.1.1
docker==4.4.1
docopt==0.6.2
elasticsearch==7.11.0
entrypoints==0.3
fake-useragent==0.1.11
filelock==3.3.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-PyMongo==2.2.0
flask-restplus==0.12.1
Flask-Script==2.0.6
fsspec==0.8.5
future==0.18.2
ga==0.1.1.dev0
gast==0.5.2
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
greenlet==1.1.2
grpcio==1.41.1
h5py==3.1.0
hdfs==2.6.0
HeapDict==1.0.1
hmmlearn==0.2.4
htmlmin==0.1.12
hupper==1.10.2
hyperopt==0.1.2
icalendar==4.0.7
idna==2.10
ImageHash==4.2.0
imageio==2.9.0
imgaug==0.4.0
immutables==0.14
importlib-metadata==4.8.1
importlib-resources==5.2.0
instructions==0.1.0
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0
joblib==1.1.0
json-tricks==3.15.3
jsonschema==4.0.0
jupyter-client==7.0.1
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
lunardate==0.2.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.0rc2
matplotlib==3.3.4
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1.1
mlflow==1.13.1
mock==4.0.3
mpmath==1.2.1
msgpack==1.0.2
msrest==0.6.21
mxnet==1.7.0.post2
nacos-sdk-python==0.1.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netCDF4==1.5.5
netifaces==0.10.9
networkx==2.5
nltk==3.5
nni==1.8
notebook==6.4.0
numba==0.53.0
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.58
packaging==21.2
pandas==1.1.5
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.2
partd==1.1.0
PasteDeploy==2.1.1
patsy==0.5.1
phik==0.11.2
pickleshare==0.7.5
Pillow==8.3.1
pip==20.2.4
pkginfo==1.7.1
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==2.4.0
prettytable==2.1.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.5
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyee==8.2.2
Pygments==2.10.0
PyJWT==2.1.0
pyluach==1.3.0
pymars==0.7.5
PyMeeus==0.5.11
pymongo==3.12.1
PyMySQL==0.10.1
pyOpenSSL==21.0.0
pyparsing==2.0.3
pyppeteer==0.2.6
pyquery==1.4.3
pyramid==1.10.4
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
PythonWebHDFS==0.2.3
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
redis==3.5.3
regex==2021.11.2
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
schema==0.7.4
scikit-image==0.17.2
scikit-learn
scipy==1.5.4
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
setuptools==58.0.4
Shapely==1.8.0
simplejson==3.17.5
six==1.16.0
smmap==3.0.5
sortedcontainers==2.4.0
soupsieve==2.3
SQLAlchemy==1.4.26
sqlparse==0.4.2
statsmodels==0.11.1
sympy==1.7.1
tabulate==0.8.9
tangled-up-in-unicode==0.2.0
tblib==1.7.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2020.9.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
translationstring==1.4
tsfresh==0.17.0
typing-extensions==3.10.0.2
tzdata==2021.5
urllib3==1.26.7
venusian==3.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
visions==0.6.0
w3lib==1.22.0
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.2.1
websockets==9.1
Werkzeug==0.14.1
wheel==0.37.0
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
wincertstore==0.2
workalendar==16.1.0
wrapt==1.12.1
xlrd==1.2.0
yapf==0.31.0
zict==2.0.0
zipp==3.6.0
zope.deprecation==4.4.0
zope.interface==5.1.2